Fix indentation check if indent starts on the same line.
In case code starts on the same line as block, indentation reports a failure if
the next lines are not aligned with the first symbol.

i.e.
void f() { int i = 3;          // OK to add the code at the end of the line
           return i * i / 3; } // Correct: aligned with previous definition

It may be a bad practice to add code on the same line as the beginning of the
block, but that should be a separate rule.