Looks like this is caused by this bug in the JDK: http://bugs.java.com/bugdatabase/view_bug.do?bug_id=7089443. The hostname is longer than 64 characters on the VM your job is running on, which is the condition for triggering this. I'm guessing that the machine you couldn't reproduce it on has a shorter hostname. All of this is probably due to the migration of our sudo-enabled Precise builds to a new infrastructure. Sorry for the inconvenience.
Supposedly this has been fixed in OpenJDK 8, but as far as I can tell was never backported into OpenJDK 7. If you can't upgrade to OpenJDK 8 I'd recommend changing the hostname before running your tests with "sudo hostname a-new-hostname". You could also just use the old hostname, but cut off the last few characters with "sudo hostname "$(hostname | cut -c1-63)"" (the bug reports are a bit unclear on whether a hostname with 64 characters works or not, so I went with 63 to be on the safe side.