Fix #775: Cannot find reports with relative Ant globbing
Upgrade Ant dependency to latest.  Newer versions of DirectoryScanner do
not require a basedir to be set if all includes are absolute, which
makes this fix much easier.

Normalize the report paths and make them all absolute.  Relative
reports/globs are made absolute using the module's base dir.

Use Ant's DirectoryScanner with the now formatted report paths and no
basedir to allow globbing/wildcards on any location on the filesystem.