Fix #766: Handle relative report paths with backticks outside basedir.
It was possible to have an NPE when normalizing backticks on relative
paths if the backticks moved the directory too far back.  This change
will handle invalid normalization of relative paths.

Normalization is reattempted against the absolute path, and if it is
still invalid then it is not given to the directory scanner, which does
not accept null elements.