make all sensors 'reportPath' properties multi-value
1. fixes #1384

Literaly all sensors derived from `CxxReportSensor` as well as `CxxSquidSensor`
use use the method `CxxReportSensor::getReports()` in order to access the
reporth paths. This method expects the `*.reportPath` property to be
a multi-value one. So fix the property access

* use `getStringArray()` method for parsing
* mark corresponding properties as multi-value

2. fixes #1376

While working on (1) I've found out, that the property
`UnitTestConfiguration.NUNIT_TEST_RESULTS_PROPERTY_KEY` is not public
accessible. This contradicts to the #1376.

On the other hand, if there should be no soupport for NUnit,
either the property must be marked as deprecated or the NUnit support
has to be removed completely. Please review.