Partial revert of "final" commit
Some of the fields, which represent the user customization of SonarQube rules were marked as `final`.
These are however changeable settings and are injected from SonarQube (see the `@RuleProperty` annotations).
Besides the fact, that this change in not logical it also **actually** prevents the fields of being changed.

E.g. I tried to create a custom quality profile with custom regular expression for `HardcodedIpCheck`.
The associated C++ project doesn't react on the customized regexp. The default one is used instead.

This reverts commit b92e83b1a13189395832da25a7db0ee16602e37e.