fix coverage on SonarCloud
Importing JaCoCo coverage report in XML format

Version 5.12 of our SonarJava analyzer deprecated use JaCoCo’s binary format (.exec files) to import coverage. This binary format is internal to the JaCoCo project, and as such there are no guarantees for backward compatibility, so it should not be used for integration purposes. As a replacement the sonar-jacoco 273 plugin, which imports JaCoCo’s XML coverage report, and this is the preferred option now.

In this guide, I will describe how to import this XML report in some common scenarios.
https://community.sonarsource.com/t/coverage-test-data-importing-jacoco-coverage-report-in-xml-format/12151