Analyze only files of 'JSON Compilation Database'
In case of 'sonar.cxx.jsonCompilationDatabase' is used, limit the analyzed files to the files contained in the 'JSON Compilation Database' file. The sensor is an intersection of the files configured via 'sonar.projectBaseDir' and the files contained in the 'JSON Compilation Database' file.

In the past, all files configured via 'sonar.projectBaseDir' were always analyzed and only the configuration (INCLUDES, MACROS) was read from the 'JSON Compilation Database'.

- filter 'sonar.sources' files
- add a JSON Compilation Database support integration test